Migrate more compiler flag probing to Charmonizer.
When possible, move probing for compiler flags which are used on core
source files out of the Perl build and into Charmonizer so that the
probing logic can be reused across other host builds.

Sometimes flags need to be duplicated rather than migrated -- for
instance, the "/TP" to make MSVC compile as C++ is still mandatory
because of our headers (though not because of the source content that
goes into $module.xs).